+++ crates/optimizer/Cargo.toml
@@ -11,6 +11,7 @@ version.workspace = true
 frick_instructions.workspace = true
 frick_operations.workspace = true
 frick_utils = { workspace = true, features = ["insert_or_push", "iter_ext"] }
+ron.workspace = true
 serde = { workspace = true, default-features = false, features = [
     "alloc",
     "derive"
