+++ crates/optimizer/src/lib.rs
@@ -1,86 +1,21 @@
 #![cfg_attr(docsrs, feature(doc_cfg))]
 #![no_std]
 
-extern crate alloc;
-
-mod inner;
-
 use alloc::vec::Vec;
 
-use frick_instructions::{BrainInstruction, ToInstructions};
+use frick_instructions::BrainInstruction;
 use frick_operations::BrainOperation;
-use frick_utils::IntoIteratorExt as _;
-use serde::{Deserialize, Serialize};
-use tracing::info;
 
-use self::inner::{passes, run_loop_pass, run_peephole_pass};
+extern crate alloc;
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[repr(transparent)]
-pub struct Optimizer {
-	ops: Vec<BrainOperation>,
-}
+mod instrs;
+mod ops;
+
+#[derive(Debug, Clone, Copy)]
+pub struct Optimizer;
 
 impl Optimizer {
-	pub fn new(ops: impl IntoIterator<Item = BrainOperation>) -> Self {
-		Self {
-			ops: ops.collect_to(),
-		}
-	}
-
-	#[tracing::instrument("optimize operations", skip(self))]
-	pub fn run(&mut self) -> Vec<BrainInstruction> {
-		let mut iteration = 0;
-
-		let mut progress = self.run_passes(iteration);
-
-		while progress {
-			iteration += 1;
-			progress = self.run_passes(iteration);
-		}
-
-		info!(iterations = iteration);
-
-		self.to_instructions()
-	}
-
-	#[tracing::instrument(skip(self))]
-	fn run_passes(&mut self, iteration: usize) -> bool {
-		let mut progress = false;
-
-		self.run_each_pass(&mut progress);
-
-		progress
-	}
-
-	fn run_each_pass(&mut self, progress: &mut bool) {
-		*progress |= run_peephole_pass(self.ops_mut(), passes::optimize_consecutive_instructions);
-
-		*progress |= run_peephole_pass(self.ops_mut(), passes::optimize_set_cell_instruction);
-		*progress |= run_loop_pass(self.ops_mut(), passes::optimize_clear_cell_instruction);
-
-		*progress |= passes::fix_beginning_instructions(self.ops_mut());
-
-		*progress |= run_peephole_pass(self.ops_mut(), passes::remove_unreachable_loops);
-
-		*progress |= run_peephole_pass(self.ops_mut(), passes::remove_comments);
-	}
-
-	#[must_use]
-	pub const fn ops(&self) -> &Vec<BrainOperation> {
-		&self.ops
-	}
-
-	pub const fn ops_mut(&mut self) -> &mut Vec<BrainOperation> {
-		&mut self.ops
-	}
-}
-
-impl ToInstructions for Optimizer {
-	fn to_instructions(&self) -> Vec<BrainInstruction> {
-		self.ops
-			.iter()
-			.flat_map(ToInstructions::to_instructions)
-			.collect()
+	pub fn run(ops: impl IntoIterator<Item = BrainOperation>) -> Vec<BrainInstruction> {
+		todo!()
 	}
 }
